Написала и задействовала три валидатора

+++ webapp/views.py
@@ -1,17 +1,17 @@
-from django.shortcuts import get_object_or_404, redirect
+from django.shortcuts import get_object_or_404, redirect, render
 from django.views import View
 from django.views.generic import TemplateView
 from webapp.models import Task
 from webapp.forms import TaskForm
 
 
-class TaskListView(TemplateView):
-    template_name = 'task_list.html'
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['tasks'] = Task.objects.all().order_by('-created_at')
-        return context
+class TaskListView(View):
+    def get(self, request, *args, **kwargs):
+        tasks = Task.objects.all()
+        context = {
+            'tasks': tasks
+        }
+        return render(request, 'task_list.html', context)
 
 
 class TaskDetailView(TemplateView):
@@ -27,57 +27,61 @@ class TaskDetailView(TemplateView):
 class TaskCreateView(View):
     template_name = 'task_form.html'
 
-    def get(self, request):
+    def get(self, request, *args, **kwargs):
         form = TaskForm()
-        return self.render(form)
+        return render(request, self.template_name, {'form': form, 'title': 'Создать задачу'})
 
-    def post(self, request):
-        form = TaskForm(request.POST)
+    def post(self, request, *args, **kwargs):
+        form = TaskForm(data=request.POST)
         if form.is_valid():
-            form.save()
+            task = Task()
+            task.summary = form.cleaned_data.get('summary')
+            task.description = form.cleaned_data.get('description')
+            task.status = form.cleaned_data.get('status')
+            task.save()
+            task.task_type.set(form.cleaned_data.get('task_type'))
             return redirect('task_list')
-        return self.render(form)
-
-    def render(self, form):
-        from django.shortcuts import render
-        return render(self.request, self.template_name, {'form': form, 'title': 'Создать задачу'})
+        return render(request, self.template_name, {'form': form, 'title': 'Создать задачу'})
 
 
-class TaskUpdateView(View):
-    template_name = 'task_form.html'
+def task_update_view(request, pk, *args, **kwargs):
+    task = get_object_or_404(Task, pk=pk)
 
-    def get(self, request, pk):
-        task = get_object_or_404(Task, pk=pk)
-        form = TaskForm(instance=task)
-        return self.render(form, task)
-
-    def post(self, request, pk):
-        task = get_object_or_404(Task, pk=pk)
-        form = TaskForm(request.POST, instance=task)
-        if form.is_valid():
-            form.save()
-            return redirect('task_detail', pk=task.pk)
-        return self.render(form, task)
-
-    def render(self, form, task):
-        from django.shortcuts import render
+    if request.method == "GET":
+        form = TaskForm(initial={
+            'summary': task.summary,
+            'description': task.description,
+            'status': task.status,
+            'task_type': task.task_type.all(),
+        })
         context = {
             'form': form,
             'task': task,
             'title': 'Редактировать задачу',
         }
-        return render(self.request, self.template_name, context)
+        return render(request, 'task_form.html', context)
+
+    elif request.method == "POST":
+        form = TaskForm(data=request.POST)
+        if form.is_valid():
+            task.summary = form.cleaned_data.get('summary')
+            task.description = form.cleaned_data.get('description')
+            task.status = form.cleaned_data.get('status')
+            task.save()
+            task.task_type.set(form.cleaned_data.get('task_type'))
+            return redirect('task_detail', pk=task.pk)
+
+        context = {
+            'form': form,
+            'task': task,
+            'title': 'Редактировать задачу',
+        }
+        return render(request, 'task_form.html', context)
 
 
-class TaskDeleteView(View):
-    template_name = 'task_confirm_delete.html'
-
-    def get(self, request, pk):
-        task = get_object_or_404(Task, pk=pk)
-        from django.shortcuts import render
-        return render(request, self.template_name, {'task': task})
-
-    def post(self, request, pk):
-        task = get_object_or_404(Task, pk=pk)
-        task.delete()
-        return redirect('task_list')
+def task_delete_view(request, *args, pk, **kwargs):
+    task = get_object_or_404(Task, pk=pk)
+    if request.method == "GET":
+        return render(request, 'task_confirm_delete.html', {'task': task})
+    task.delete()
+    return redirect('task_list')
